apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
asyncpg==0.30.0
prisma==0.15.0
httpx==0.28.1

# AI / LLM
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
openai==1.59.6
tiktoken==0.8.0
numpy==2.2.1

# Scraping
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Email ingestion
aiosmtpd==1.4.6

# Queue / Workers
celery==5.4.0
redis==5.2.1
kombu==5.4.2

# AWS
boto3==1.35.91
botocore==1.35.91

# Email notifications
resend==2.6.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
